Initial support for numeric types in schemas (#485)
This change does not change the enumeration type to map to any of the
new types, but leaves it still implicitly mapped to fmi3Int32 (the successor
of the original fmi3Integer type).  Making the enumeration type map to
any of the new integer types would be useful addition, but would make
code quite a bit more complex, and also leave the schema definitions not
as expressive, since the valid ranges for min/max, etc. will depend on
an attribute.

* Parse the XSD schema files in CI (#486)