Add extra directory for meta-data transport
This proposal mirrors the way SSP defines extra directory for additional
meta-data that is supposed to travel with the FMU/SSP, and can be added,
modified, or deleted by tools other than the FMU creator (in addition to
the FMU creator).  This mechanism allows for layered standards
specifying the exact nature of added meta-data, and other extensions.