Add extra directory for metadata transport (#586)
This proposal mirrors the way SSP defines extra directory for additional
metadata that is supposed to travel with the FMU, and can be added,
modified, or deleted by tools other than the FMU creator (in addition to
the FMU creator).  This mechanism allows for layered standards
specifying the exact nature of added meta-data, and other extensions.