Rewrite equation for adjoint derivative.
The text a few lines above defines the adjoint derivative as a
transposed vector multiplied from left to the jacobian. But the equation
here defines it as a vector multiplied from right to the transposed
jacobian. Both represenations are mathematically equally but for
convinience the text and the equation should be analog.
I see the form transposed vector multiplied from left to the jacobian
the more natrual one. That's why I switch to this form consistently.