Clarify the row-major serialization.
The row-major serialization is always done on a dense view of the
arrays.
Hence, the dependencies given by fmi3GetVariableDependencies does not
influence the row-major serialization.
This may lead to many 0 values in setter/getter API calls for sparse
arrays.
But everything else will be quite complicated and even not possible for
the values in the XML file which are also serialized using row-major
format since the per element dependencies are not known statically in
the XML file.