Fixed the description of causality indpendent.
Removed "All variables are a function of this <<independent>> variable."
since this is not true. E.g. a constant variable is not a function of
the independent.
Moreover this will implicitly mean that in ModelStructure this
independent variable is always a dependent variable of each other variable.

Removed the text that an implicilty defined independent variable has the
name = `time`. Since the name of all variables must be unique this would
mean that we need to disallow name = `time` of all other variables when
no independent variable is defined explicitly.