Clarify the number of states/event indicators
Getting the number of states and event indicators cannot be done
consistently since every state is also present as a variable but event
indicators are not.
Now a clear defintion is given how the static number of states/events
can be get without using C-API function calls and also how to get the
dependent structural parameter the number of states/events depends on.