Define event indicators as variables (#587)
The current FMI 3.0 has a well defined variable (including a
valueReference) for all types of values, including states, state
derivatives and the independent variable (time), but not for event
indicators.
This commit fixes this.
Event indicators are listed now in ModelStructure and have therefore a
corresponding variable with a valueReference. Event indicators can be of
causality local and output (as states).
The c-API function for getting event indicators keep the same but
fmi3GetFloat64 can now also be used.