CI adaptions
Create a docker image usable to build the examples.
Use this docker image for the example build in CI.
Added .sh/.bat scripts to build the doc and examples locally (with docker installed).
See CONTRIBUTING.adoc for details of the local build.

Required adaptions before a merge of this commit:
- Create a modelica or FMI organization on DockerHub (analog to the modelica organization on GitHub).
- Push the docker image build_examples to this DockerHub organization.
- Adapt the scripts to use this docker image (for now my DockerHub account "fm12" is used)
- Maybe automatically rebuild the docker image by CI and push it to DockerHub
  (The DockerHub login credentials needs to be configure in CI for this.)
  But the image should rarely be changed, so a pure manual update of the image is also possisble.