Minimal changes to fully support clock arrays
This changeset includes the minimal changes necessary to fully allow the
use of arrays of clock variables. Further refactoring could be done
based on this to merge the Set/GetClock functions into the getter and
setter sections of the document, and other cleanups might be wanted, but
this should demonstrate what is basically needed to do this.

The special treatment of index 0 follows the convention used for
dependency information, but could be dropped if not wanted.