Fix enumeration variable definition (#1034)
Using a restriction to remove the declaredType of the base to then
re-add it is problematic, because a restriction will remove all complex
content that is not re-iterated, thus all content from the base not to
be removed needs to be restated, which would be a maintenance nightmare.

This restructures the type hierarchy by only introducing the
declaredType attribute in the relevant separated base elements.